NotificationsSection component for page and popup
