+++ apps/app/src/components/NotificationsSection/NotificationsSection.module.scss
@@ -1,0 +1,27 @@
+.notificationsContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+  margin-top: 35px;
+}
+
+.noNotificationsContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 85%;
+  margin: 0 30px;
+}
+
+.noNotificationsTitle {
+  @include heading-3;
+  color: $black-50;
+  margin-top: 1.5rem;
+}
+
+.noNotificationsLabel {
+  @include paragraph-14;
+  color: $black-50;
+  margin-top: 6px;
+}
