+++ apps/app/src/components/NotificationsSection/index.ts
@@ -1,0 +1,2 @@
+import NotificationsSection from './NotificationsSection';
+export default NotificationsSection;
